Add new song: Temporal Coupling
Temporal Coupling is a song that works hard to establish a sound space. Creating layers sonically, synchronizing harmonies and rhythms effortlessly. It's one of those songs that can be repeated endlessly with no sign of ending. The Empty Language's first purely instrumental track is one of the best on the album.